changed price column in db from integer to decimal value type, got homepage info needed

+++ backend/db/migrations/20210308194210-create-product.js
@@ -22,7 +22,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL
       },
       imageLink: {
         allowNull: false,
